[Feature] 주소의 경도와 위도로 x y좌표 변환 후 위치 저장

+++ src/main/java/com/overcomingroom/bellbell/weather/domain/dto/WeatherAndClothesDto.java
@@ -14,7 +14,7 @@ public class WeatherAndClothesDto {
     private LocalTime now; // 현재 시각
     private String location; // 발표 지역
     private String weather;
-    private Integer temp;
+    private Float temp;
     private String clothes;
 
     public static String weatherAndClothesInfo(WeatherAndClothesDto weatherAndClothesDto) {
